[vmware] add support for disconnecting a volume via HttpNfc upload
os-brick needs to be able to upload a vmdk file via HttpNfc upload
after obtaining a lease from ImportVApp operation. This will be
used in a situation like restoring a VVol volume from a backup.
os-brick will receive information from cinder-volume of how to
construct the ImportVApp call to VMware API, via import_data.
This is backward compatible with the initial method of uploading
the vmdk to VMWare.